renamed package "numpy" back to "numpy_utils"
