+++ mathics/builtin/randomnumbers.py
@@ -20,7 +20,7 @@ from operator import mul as operator_mul
 from functools import reduce
 
 from mathics.builtin.base import Builtin
-from mathics.builtin.numpy import instantiate_elements, stack
+from mathics.builtin.numpy_utils import instantiate_elements, stack
 from mathics.core.expression import (Integer, String, Symbol, Real, Expression,
                                      Complex)
