+++ test/test_color.py
@@ -10,7 +10,7 @@ from random import random
 from six.moves import range
 
 import mathics.builtin.colors as colors
-from mathics.builtin.numpy.with_numpy import array, vectorized
+from mathics.builtin.numpy_utils import array, vectorized
 from mathics.core.definitions import Definitions
 from mathics.core.evaluation import Evaluation
 from mathics.core.expression import Expression
